nomic_indexer/requirements.txt
torch==2.1.0
sentence-transformers==2.6.1
faiss-cpu==1.7.4
flask==2.2.5
flask-cors==3.0.10
numpy==1.26.4
